Commit time refactor (#142)
* initial commit with sample code

* Create gitlabcommitcollector.py

* class breakout

* bitbucket notes

* Almost working

* removing extra files

* Don't need to pass namespace

* latest

* latest from working session

* working version

* it works!

* change type to provider, set default

* Update docs

* Updating new env vars, add support for legacy vars

* Rename env vars, deprecate old ones

* Fix legacy var support

* Mark deprecated vars as not req

* Add blurb for supported git providers

* Move default value to pelorus.py

* Remove git provider reference in docs

Only one is supported

* Update language

* Add unsupported messaging

* Fix

* fix
