+++ exporters/pelorus.py
@@ -1,3 +1,4 @@
+from abc import ABC
 import logging
 import os
 import sys
@@ -9,6 +10,7 @@ DEFAULT_PROD_LABEL = ''
 DEFAULT_LOG_LEVEL = 'INFO'
 DEFAULT_LOG_FORMAT = '%(asctime)-15s %(levelname)-8s %(message)s'
 DEFAULT_LOG_DATE_FORMAT = '%m-%d-%Y %H:%M:%S'
+DEFAULT_GIT = "github"
 
 loglevel = os.getenv('LOG_LEVEL', DEFAULT_LOG_LEVEL)
 numeric_level = getattr(logging, loglevel.upper(), None)
@@ -57,3 +59,23 @@ def check_required_config(vars):
     if missing_configs:
         logging.error("This program will exit.")
         sys.exit(1)
+
+
+def check_legacy_vars():
+    username = os.environ.get('GITHUB_USER')
+    token = os.environ.get('GITHUB_TOKEN')
+    api = os.environ.get('GITHUB_API')
+    if username is not None:
+        os.environ['GIT_USER'] = username
+    if token is not None:
+        os.environ['GIT_TOKEN'] = token
+    if api is not None:
+        os.environ['GIT_API'] = api
+
+
+class AbstractPelorusExporter(ABC):
+    """
+    Base class for PelorusExporter
+    """
+    def __init_():
+        pass
